+++ resources/measures/HPXMLtoOpenStudio/resources/airflow.rb
@@ -39,6 +39,9 @@ class Airflow
       runner.registerError("Cannot determine the number of above grade stories.")
       return false
     end
+    building.crawlspace = []
+    building.unfinished_basement = []
+    building.unfinished_attic = []
     building.stories = model.getBuilding.standardsNumberOfAboveGroundStories.get
 
     building.ag_ext_wall_area = 0
@@ -46,33 +49,41 @@ class Airflow
     building.above_grade_volume = 0
     units = Geometry.get_building_units(model, runner)
     units.each do |unit|
-      units_represented = 1
-      if unit.additionalProperties.getFeatureAsInteger("Units Represented").is_initialized
-        units_represented = unit.additionalProperties.getFeatureAsInteger("Units Represented").get
-      end
-
-      building.above_grade_volume += units_represented * Geometry.get_above_grade_finished_volume_from_spaces(unit.spaces, runner)
-      building.ag_ext_wall_area += units_represented * Geometry.calculate_above_grade_exterior_wall_area(unit.spaces)
-      building.ag_ffa += units_represented * Geometry.get_above_grade_finished_floor_area_from_spaces(unit.spaces, runner)
+      building.above_grade_volume += Geometry.get_above_grade_finished_volume_from_spaces(unit.spaces, runner)
+      building.ag_ext_wall_area += Geometry.calculate_above_grade_exterior_wall_area(unit.spaces)
+      building.ag_ffa += Geometry.get_above_grade_finished_floor_area_from_spaces(unit.spaces, runner)
     end
 
     model.getThermalZones.each do |thermal_zone|
       if Geometry.is_garage(thermal_zone)
         building.garage = ZoneInfo.new(thermal_zone, Geometry.get_height_of_spaces(thermal_zone.spaces), UnitConversions.convert(thermal_zone.floorArea, "m^2", "ft^2"), Geometry.get_zone_volume(thermal_zone, runner), Geometry.get_z_origin_for_zone(thermal_zone), nil, nil)
       elsif Geometry.is_unfinished_basement(thermal_zone)
-        building.unfinished_basement = ZoneInfo.new(thermal_zone, Geometry.get_height_of_spaces(thermal_zone.spaces), UnitConversions.convert(thermal_zone.floorArea, "m^2", "ft^2"), Geometry.get_zone_volume(thermal_zone, runner), Geometry.get_z_origin_for_zone(thermal_zone), infil.unfinished_basement_ach, nil)
+        building.unfinished_basement << ZoneInfo.new(thermal_zone, Geometry.get_height_of_spaces(thermal_zone.spaces), UnitConversions.convert(thermal_zone.floorArea, "m^2", "ft^2"), Geometry.get_zone_volume(thermal_zone, runner), Geometry.get_z_origin_for_zone(thermal_zone), infil.unfinished_basement_ach, nil)
       elsif Geometry.is_crawl(thermal_zone)
-        building.crawlspace = ZoneInfo.new(thermal_zone, Geometry.get_height_of_spaces(thermal_zone.spaces), UnitConversions.convert(thermal_zone.floorArea, "m^2", "ft^2"), Geometry.get_zone_volume(thermal_zone, runner), Geometry.get_z_origin_for_zone(thermal_zone), infil.crawl_ach, nil)
+        building.crawlspace << ZoneInfo.new(thermal_zone, Geometry.get_height_of_spaces(thermal_zone.spaces), UnitConversions.convert(thermal_zone.floorArea, "m^2", "ft^2"), Geometry.get_zone_volume(thermal_zone, runner), Geometry.get_z_origin_for_zone(thermal_zone), infil.crawl_ach, nil)
       elsif Geometry.is_pier_beam(thermal_zone)
         building.pierbeam = ZoneInfo.new(thermal_zone, Geometry.get_height_of_spaces(thermal_zone.spaces), UnitConversions.convert(thermal_zone.floorArea, "m^2", "ft^2"), Geometry.get_zone_volume(thermal_zone, runner), Geometry.get_z_origin_for_zone(thermal_zone), infil.pier_beam_ach, nil)
       elsif Geometry.is_unfinished_attic(thermal_zone)
-        building.unfinished_attic = ZoneInfo.new(thermal_zone, Geometry.get_height_of_spaces(thermal_zone.spaces), UnitConversions.convert(thermal_zone.floorArea, "m^2", "ft^2"), Geometry.get_zone_volume(thermal_zone, runner), Geometry.get_z_origin_for_zone(thermal_zone), infil.unfinished_attic_const_ach, infil.unfinished_attic_sla)
+        building.unfinished_attic << ZoneInfo.new(thermal_zone, Geometry.get_height_of_spaces(thermal_zone.spaces), UnitConversions.convert(thermal_zone.floorArea, "m^2", "ft^2"), Geometry.get_zone_volume(thermal_zone, runner), Geometry.get_z_origin_for_zone(thermal_zone), infil.unfinished_attic_const_ach, infil.unfinished_attic_sla)
       end
     end
 
     return false if building.ag_ffa.nil?
 
-    wind_speed = process_wind_speed_correction(infil.terrain, infil.shelter_coef, Geometry.get_closest_neighbor_distance(model), building.building_height)
+    building_height = nil
+    num_floors = model.getBuilding.additionalProperties.getFeatureAsInteger("num_floors")
+    if Geometry.get_building_type(model) == Constants.BuildingTypeMultifamily
+      units.each do |unit|
+        Geometry.get_thermal_zones_from_spaces(unit.spaces).each do |thermal_zone|
+          next unless Geometry.is_living(thermal_zone)
+
+          building.building_height = num_floors.get.to_f * Geometry.get_height_of_spaces(thermal_zone.spaces)
+        end
+      end
+    end
+
+    building_height = building.building_height
+    wind_speed = process_wind_speed_correction(infil.terrain, infil.shelter_coef, Geometry.get_closest_neighbor_distance(model), building_height)
     if not process_infiltration(model, infil, wind_speed, building, weather)
       return false
     end
@@ -151,8 +162,6 @@ class Airflow
       tout_sensor.setName("#{obj_name_airflow} tt s")
       tout_sensor.setKeyName(unit_living.zone.name.to_s)
 
-      # Update model
-
       success, infil_output = process_infiltration_for_unit(model, runner, obj_name_infil, infil, wind_speed, building, weather, unit_ag_ffa, unit_ag_ext_wall_area, unit_living, unit_finished_basement)
       return false if not success
 
@@ -211,17 +220,23 @@ class Airflow
     end # end unit loop
 
     # Store info for HVAC Sizing measure
-    unless building.crawlspace.nil?
-      building.crawlspace.zone.additionalProperties.setFeature(Constants.SizingInfoZoneInfiltrationCFM, building.crawlspace.inf_flow.to_f)
+    unless building.crawlspace.empty?
+      building.crawlspace.each do |cs|
+        cs.zone.additionalProperties.setFeature(Constants.SizingInfoZoneInfiltrationCFM, cs.inf_flow.to_f)
+      end
     end
     unless building.pierbeam.nil?
       building.pierbeam.zone.additionalProperties.setFeature(Constants.SizingInfoZoneInfiltrationCFM, building.pierbeam.inf_flow.to_f)
     end
-    unless building.unfinished_basement.nil?
-      building.unfinished_basement.zone.additionalProperties.setFeature(Constants.SizingInfoZoneInfiltrationCFM, building.unfinished_basement.inf_flow.to_f)
+    unless building.unfinished_basement.empty?
+      building.unfinished_basement.each do |ub|
+        ub.zone.additionalProperties.setFeature(Constants.SizingInfoZoneInfiltrationCFM, ub.inf_flow.to_f)
+      end
     end
-    unless building.unfinished_attic.nil?
-      building.unfinished_attic.zone.additionalProperties.setFeature(Constants.SizingInfoZoneInfiltrationCFM, building.unfinished_attic.inf_flow)
+    unless building.unfinished_attic.empty?
+      building.unfinished_attic.each do |ua|
+        ua.zone.additionalProperties.setFeature(Constants.SizingInfoZoneInfiltrationCFM, ua.inf_flow)
+      end
     end
 
     terrain = { Constants.TerrainOcean => "Ocean", # Ocean, Bayou flat country
@@ -491,10 +506,11 @@ class Airflow
   def self.process_infiltration(model, infil, wind_speed, building, weather)
     spaces = []
     spaces << building.garage if not building.garage.nil?
-    spaces << building.unfinished_basement if not building.unfinished_basement.nil?
-    spaces << building.crawlspace if not building.crawlspace.nil?
+    # Multiple foundation spaces
+    spaces += building.unfinished_basement if not building.unfinished_basement.empty?
+    spaces += building.crawlspace if not building.crawlspace.empty?
     spaces << building.pierbeam if not building.pierbeam.nil?
-    spaces << building.unfinished_attic if not building.unfinished_attic.nil?
+    spaces += building.unfinished_attic if not building.unfinished_attic.empty?
 
     unless building.garage.nil?
       building.garage.inf_method = @infMethodSG
@@ -505,14 +521,20 @@ class Airflow
       building.garage.inf_flow = building.garage.ACH / UnitConversions.convert(1.0, "hr", "min") * building.garage.volume # cfm
     end
 
-    unless building.unfinished_basement.nil?
-      building.unfinished_basement.inf_method = @infMethodRes # Used for constant ACH
-      building.unfinished_basement.inf_flow = building.unfinished_basement.ACH / UnitConversions.convert(1.0, "hr", "min") * building.unfinished_basement.volume
+    unless building.unfinished_basement.empty?
+      building.unfinished_basement.each do |ub|
+        # building.unfinished_basement.inf_method = @infMethodRes # Used for constant ACH
+        # building.unfinished_basement.inf_flow = building.unfinished_basement.ACH / UnitConversions.convert(1.0, "hr", "min") * building.unfinished_basement.volume
+        ub.inf_method = @infMethodRes # Used for constant ACH
+        ub.inf_flow = ub.ACH / UnitConversions.convert(1.0, "hr", "min") * ub.volume
+      end
     end
 
-    unless building.crawlspace.nil?
-      building.crawlspace.inf_method = @infMethodRes
-      building.crawlspace.inf_flow = building.crawlspace.ACH / UnitConversions.convert(1.0, "hr", "min") * building.crawlspace.volume
+    unless building.crawlspace.empty?
+      building.crawlspace.each do |cs|
+        cs.inf_method = @infMethodRes
+        cs.inf_flow = cs.ACH / UnitConversions.convert(1.0, "hr", "min") * cs.volume
+      end
     end
 
     unless building.pierbeam.nil?
@@ -520,20 +542,21 @@ class Airflow
       building.pierbeam.inf_flow = building.pierbeam.ACH / UnitConversions.convert(1.0, "hr", "min") * building.pierbeam.volume
     end
 
-    unless building.unfinished_attic.nil?
-      if not building.unfinished_attic.SLA.nil?
-        building.unfinished_attic.inf_method = @infMethodSG
-        building.unfinished_attic.hor_lk_frac = 0.75 # Same as Energy Gauge USA Attic Model
-        building.unfinished_attic.neutral_level = 0.5 # DOE-2 Default
-        building.unfinished_attic.ACH = Airflow.get_infiltration_ACH_from_SLA(building.unfinished_attic.SLA, 1.0, weather)
-      elsif not building.unfinished_attic.ACH.nil?
-        building.unfinished_attic.inf_method = @infMethodRes
+    unless building.unfinished_attic.empty?
+      building.unfinished_attic.each do |ua|
+        if not ua.SLA.nil?
+          ua.inf_method = @infMethodSG
+          ua.hor_lk_frac = 0.75 # Same as Energy Gauge USA Attic Model
+          ua.neutral_level = 0.5 # DOE-2 Default
+          ua.ACH = Airflow.get_infiltration_ACH_from_SLA(ua.SLA, 1.0, weather)
+        elsif not ua.ACH.nil?
+          ua.inf_method = @infMethodRes
+        end
+        ua.inf_flow = ua.ACH / UnitConversions.convert(1.0, "hr", "min") * ua.volume
       end
-      building.unfinished_attic.inf_flow = building.unfinished_attic.ACH / UnitConversions.convert(1.0, "hr", "min") * building.unfinished_attic.volume
     end
 
     process_infiltration_for_spaces(model, spaces, wind_speed)
-
     return true
   end
 
@@ -559,8 +582,78 @@ class Airflow
       # Calculate SLA for above-grade portion of the building
       building.SLA = Airflow.get_infiltration_SLA_from_ACH50(infil.living_ach50, n_i, building.ag_ffa, building.above_grade_volume)
 
-      # Effective Leakage Area (ft^2)
-      a_o = building.SLA * building.ag_ffa * (unit_ag_ext_wall_area / building.ag_ext_wall_area)
+      # Calculate unit ELA proportional to exposed exterior wall
+      n_units = model.getBuilding.additionalProperties.getFeatureAsInteger("num_units")
+      has_rear_units = model.getBuilding.additionalProperties.getFeatureAsBoolean("has_rear_units")
+      num_floors = model.getBuilding.additionalProperties.getFeatureAsInteger("num_floors")
+      if n_units.is_initialized
+        n_units = n_units.get
+      end
+      if has_rear_units.is_initialized
+        has_rear_units = has_rear_units.get
+      end
+      if num_floors.is_initialized
+        num_floors = num_floors.get.to_f
+      end
+
+      building_type = Geometry.get_building_type(model)
+      # Infiltration for MF/SFA
+      if building_type == Constants.BuildingTypeMultifamily or building_type == Constants.BuildingTypeSingleFamilyAttached
+        facade_areas = {}
+
+        model.getThermalZones.each do |thermal_zone|
+          next unless thermal_zone.name.to_s.start_with? "living"
+
+          thermal_zone.spaces.each do |space|
+            space.surfaces.each do |surface|
+              next if surface.surfaceType.downcase != "wall"
+              next if surface.outsideBoundaryCondition.downcase == "foundation"
+
+              if not facade_areas[Geometry.get_facade_for_surface(surface)]
+                facade_areas[Geometry.get_facade_for_surface(surface)] = 0
+              end
+              facade_areas[Geometry.get_facade_for_surface(surface)] += surface.grossArea
+            end
+          end
+        end
+
+        building_ag_ffa = unit_ag_ffa * n_units
+        mf_building_ELA = building_ag_ffa * building.SLA
+
+        if Geometry.get_building_type(model) == Constants.BuildingTypeMultifamily
+          num_units_per_floor = n_units / num_floors
+        elsif Geometry.get_building_type(model) == Constants.BuildingTypeSingleFamilyAttached
+          num_floors = 1.0
+          num_units_per_floor = n_units
+        end
+
+        if num_units_per_floor == 1 or num_units_per_floor == 2 or (num_units_per_floor == 4 and has_rear_units) # No middle unit(s)
+          a_o_frac = 1 / num_floors / num_units_per_floor # all units have same exterior wall area
+          a_o = mf_building_ELA * a_o_frac
+        else # Has middle unit(s)
+          if has_rear_units
+            n_end_units = 4 * num_floors
+            n_mid_units = n_units - n_end_units
+            ext_wall_area_end = facade_areas["front"] + facade_areas["left"]
+            ext_wall_area_mid = facade_areas["front"]
+            ext_wall_area_building = (n_end_units * ext_wall_area_end) + (n_mid_units * ext_wall_area_mid)
+            ext_wall_area_building = UnitConversions.convert(ext_wall_area_building, "m^2", "ft^2")
+          else
+            n_end_units = 2 * num_floors
+            n_mid_units = n_units - n_end_units
+            ext_wall_area_end = facade_areas["front"] + facade_areas["left"] + facade_areas["back"]
+            ext_wall_area_mid = facade_areas["front"] + facade_areas["back"]
+            ext_wall_area_building = (n_end_units * ext_wall_area_end) + (n_mid_units * ext_wall_area_mid)
+            ext_wall_area_building = UnitConversions.convert(ext_wall_area_building, "m^2", "ft^2")
+          end
+          a_o = building.SLA * building_ag_ffa * (unit_ag_ext_wall_area / ext_wall_area_building) # Effective Leakage Area (ft^2) - Unit
+        end
+      else # SFD
+        num_floors = building.stories
+        a_o = building.SLA * building.ag_ffa * (unit_ag_ext_wall_area / building.ag_ext_wall_area) # Effective Leakage Area (ft^2) - Unit
+        building_ag_ffa = building.ag_ffa
+        ext_wall_area_building = building.ag_ext_wall_area
+      end
 
       # Calculate SLA for unit
       unit_living.SLA = a_o / unit_ag_ffa
@@ -569,8 +662,12 @@ class Airflow
       c_i = a_o * (2.0 / outside_air_density)**0.5 * delta_pref**(0.5 - n_i) * inf_conv_factor
 
       if infil.has_flue_chimney
-        y_i = 0.2 # Fraction of leakage through the flue; 0.2 is a "typical" value according to THE ALBERTA AIR INFIL1RATION MODEL, Walker and Wilson, 1990
-        flue_height = building.building_height + 2.0 # ft
+        y_i = 0.2 # Fraction of leakage through the flue; 0.2 is a "typical" value according to THE ALBERTA AIR INFILTRATION MODEL, Walker and Wilson, 1990
+        if building_type == Constants.BuildingTypeMultifamily
+          flue_height = unit_living.height * num_floors + 2.0 # ft
+        else
+          flue_height = building.building_height + 2.0 # ft
+        end
         s_wflue = 1.0 # Flue Shelter Coefficient
       else
         y_i = 0.0 # Fraction of leakage through the flu
@@ -579,28 +676,45 @@ class Airflow
       end
 
       vented_crawl = false
-      if (not building.crawlspace.nil? and building.crawlspace.ACH > 0) or (not building.pierbeam.nil? and building.pierbeam.ACH > 0)
+      if not building.crawlspace.empty?
+        building.crawlspace.each do |cs|
+          if cs.ACH > 0
+            vented_crawl = true
+          end
+        end
+      end
+
+      if (not building.pierbeam.nil? and building.pierbeam.ACH > 0)
         vented_crawl = true
+      end
+
+      # MF vented_crawl
+      found_type = model.getBuilding.additionalProperties.getFeatureAsString("found_type")
+      if found_type.is_initialized
+        found_type = found_type.get
+        if (infil.crawl_ach > 0.0 and found_type == "crawlspace") or (infil.pier_beam_ach > 0 and found_type == "pier and beam")
+          vented_crawl = true
+        end
       end
 
       # Leakage distributions per Iain Walker (LBL) recommendations
       if vented_crawl
         # 15% ceiling, 35% walls, 50% floor leakage distribution for vented crawl
-        leakkage_ceiling = 0.15
+        leakage_ceiling = 0.15
         leakage_walls = 0.35
         leakage_floor = 0.50
       else
         # 25% ceiling, 50% walls, 25% floor leakage distribution for slab/basement/unvented crawl
-        leakkage_ceiling = 0.25
+        leakage_ceiling = 0.25
         leakage_walls = 0.50
         leakage_floor = 0.25
       end
-      if leakkage_ceiling + leakage_walls + leakage_floor != 1
-        runner.registerError("Invalid air leakage distribution specified (#{leakkage_ceiling}, #{leakage_walls}, #{leakage_floor}); does not add up to 1.")
+      if leakage_ceiling + leakage_walls + leakage_floor != 1
+        runner.registerError("Invalid air leakage distribution specified (#{leakage_ceiling}, #{leakage_walls}, #{leakage_floor}); does not add up to 1.")
         return false
       end
-      r_i = (leakkage_ceiling + leakage_floor)
-      x_i = (leakkage_ceiling - leakage_floor)
+      r_i = (leakage_ceiling + leakage_floor)
+      x_i = (leakage_ceiling - leakage_floor)
       r_i = r_i * (1 - y_i)
       x_i = x_i * (1 - y_i)
 
@@ -660,8 +774,7 @@ class Airflow
       end
 
       wind_coef = f_w * UnitConversions.convert(outside_air_density / 2.0, "lbm/ft^3", "inH2O/mph^2")**n_i # inH2O^n/mph^2n
-
-      unit_living.ACH = Airflow.get_infiltration_ACH_from_SLA(unit_living.SLA, building.stories, weather)
+      unit_living.ACH = Airflow.get_infiltration_ACH_from_SLA(unit_living.SLA, num_floors, weather)
 
       # Convert living space ACH to cfm:
       unit_living.inf_flow = unit_living.ACH / UnitConversions.convert(1.0, "hr", "min") * unit_living.volume # cfm
@@ -672,7 +785,6 @@ class Airflow
 
       unit_living.ACH = infil.living_constant_ach
       unit_living.inf_flow = unit_living.ACH / UnitConversions.convert(1.0, "hr", "min") * unit_living.volume # cfm
-
     end
 
     unless unit_finished_basement.nil?
@@ -1148,14 +1260,37 @@ class Airflow
         frac_oa = 0
       elsif not unit_finished_basement.nil? and unit_finished_basement.zone == location_zone
         frac_oa = 0
-      elsif not building.unfinished_basement.nil? and building.unfinished_basement.zone == location_zone
-        frac_oa = 0
-      elsif not building.crawlspace.nil? and building.crawlspace.zone == location_zone and building.crawlspace.ACH == 0
-        frac_oa = 0
+      elsif not building.unfinished_basement.empty?
+        building.unfinished_basement.each do |ub|
+          if ub.zone == location_zone
+            frac_oa = 0
+            break
+          else
+            frac_oa = 1
+          end
+        end
+      elsif not building.crawlspace.empty?
+        building.crawlspace.each do |cs|
+          if cs.zone == location_zone and cs.ACH == 0
+            frac_oa = 0
+            break
+          else
+            frac_oa = 1
+          end
+        end
       elsif not building.pierbeam.nil? and building.pierbeam.zone == location_zone and building.pierbeam.ACH == 0
         frac_oa = 0
-      elsif not building.unfinished_attic.nil? and building.unfinished_attic.zone == location_zone and building.unfinished_attic.ACH == 0
-        frac_oa = 0
+      # elsif not building.unfinished_attic.nil? and building.unfinished_attic.zone == location_zone and building.unfinished_attic.ACH == 0
+      #   frac_oa = 0
+      elsif not building.unfinished_attic.empty?
+        building.unfinished_attic.each do |ua|
+          if ua.zone == location_zone and ua.ACH == 0
+            frac_oa = 0
+            break
+          else
+            frac_oa = 1
+          end
+        end
       else
         # Assume that all of the unbalanced make-up air is driven infiltration from outdoors.
         # This assumes that the holes for attic ventilation are much larger than any attic bypasses.