+++ resources/measures/ResidentialConstructionsSlab/measure.rb
@@ -126,6 +126,23 @@ class ProcessConstructionsSlab < OpenStudio::Measure::ModelMeasure
 
     # Get geometry values
     surfaces = floors_by_type[Constants.SurfaceTypeFloorFndGrndFinSlab]
+    space_surfaces = []
+    living_space = nil
+    surfaces.each do |surface|
+      next if not Geometry.is_living(surface.space.get)
+
+      living_space = surface.space.get
+      break
+    end
+
+    surfaces.each do |surface|
+      if surface.space.get == living_space
+        space_surfaces << surface
+      end
+    end
+
+    # Calculate slab area based on one unit
+    slabArea = Geometry.calculate_total_area_from_surfaces(space_surfaces)
 
     # Define materials
     slabCarpetPerimeterConduction, slabBarePerimeterConduction = SlabPerimeterConductancesByType(perimeter_r, gap_r, perimeter_width, exterior_r, whole_r, exterior_depth)
@@ -133,9 +150,8 @@ class ProcessConstructionsSlab < OpenStudio::Measure::ModelMeasure
     slab_perimeter_conduction = slabCarpetPerimeterConduction * carpetFloorFraction + slabBarePerimeterConduction * (1 - carpetFloorFraction)
 
     surfaces.each do |surface|
-      slabArea = UnitConversions.convert(surface.grossArea, "m^2", "ft^2")
+      # slabArea = UnitConversions.convert(surface.grossArea, "m^2", "ft^2")
       slabExtPerimeter = Geometry.calculate_exposed_perimeter(model, [surface], false)
-
       if slabExtPerimeter > 0
         effective_slab_Rvalue = slabArea / (slabExtPerimeter * slab_perimeter_conduction)
       else