+++ resources/measures/ResidentialConstructionsWallsDoubleWoodStud/tests/DoubleWoodStud_Test.rb
@@ -106,7 +106,7 @@ class ProcessConstructionsWallsDoubleWoodStudTest < MiniTest::Test
     args_hash["framing_spacing"] = 24
     args_hash["is_staggered"] = false
     expected_num_del_objects = {}
-    expected_num_new_objects = { "Material" => 9, "Construction" => 9, "InternalMass" => 8, "InternalMassDefinition" => 8 }
+    expected_num_new_objects = { "Material" => 10, "Construction" => 6, "InternalMass" => 2, "InternalMassDefinition" => 2 }
     ext_finish_r = 0.009525 / 0.089435
     osb_r = 0.0127 / 0.1154577
     stud_cavity_r = 0.0889 / 0.074106442682155
@@ -127,7 +127,7 @@ class ProcessConstructionsWallsDoubleWoodStudTest < MiniTest::Test
     args_hash["framing_spacing"] = 24
     args_hash["is_staggered"] = false
     expected_num_del_objects = {}
-    expected_num_new_objects = { "Material" => 10, "Construction" => 14, "InternalMass" => 16, "InternalMassDefinition" => 16 }
+    expected_num_new_objects = { "Material" => 10, "Construction" => 6, "InternalMass" => 2, "InternalMassDefinition" => 2 }
     ext_finish_r = 0.009525 / 0.089435
     osb_r = 0.0127 / 0.1154577
     stud_cavity_r = 0.0889 / 0.074106442682155